+++ src/main/java/frc/robot/subsystems/drive/Drive.java
@@ -4,33 +4,16 @@ import edu.wpi.first.math.Matrix;
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.numbers.N1;
 import edu.wpi.first.math.numbers.N3;
-import edu.wpi.first.math.util.Units;
-import edu.wpi.first.wpilibj.XboxController;
 import frc.robot.pioneersLib.bumSwerve.SwerveDrive;
 import frc.robot.pioneersLib.bumSwerve.SwerveModule;
 import frc.robot.pioneersLib.bumSwerve.Gyro.SwerveGyroIO;
-import frc.robot.pioneersLib.bumSwerve.Gyro.SwerveGyroIONavX;
-import frc.robot.pioneersLib.bumSwerve.Gyro.SwerveGyroIOSim;
-import frc.robot.pioneersLib.bumSwerve.SwerveAbsoluteEncoder.SwerveAbsoluteEncoderIOCANcoder;
-import frc.robot.pioneersLib.bumSwerve.SwerveAbsoluteEncoder.SwerveAbsoluteEncoderIOSim;
-import frc.robot.pioneersLib.bumSwerve.SwerveMotor.SwerveMotorIOKrakenSim;
-import frc.robot.pioneersLib.bumSwerve.SwerveMotor.SwerveMotorIONeoSim;
-import frc.robot.pioneersLib.bumSwerve.SwerveMotor.SwerveMotorIOSparkMax;
-import frc.robot.pioneersLib.bumSwerve.SwerveMotor.SwerveMotorIOTalonFX;
 import java.util.function.DoubleSupplier;
 
 import frc.robot.Constants;
-import frc.robot.pioneersLib.bumSwerve.SwerveDrive;
-import frc.robot.pioneersLib.bumSwerve.SwerveModule;
-import frc.robot.pioneersLib.bumSwerve.Gyro.SwerveGyroIO;
 import frc.robot.pioneersLib.subsystem.Subsystem;
-import frc.robot.subsystems.vision.Vision;
-import frc.robot.subsystems.vision.VisionIOReal;
-import frc.robot.subsystems.vision.VisionIOSim;
 
 public class Drive extends Subsystem<DriveStates> {
     private SwerveDrive drive;
-    private Vision vision;
 
     public Drive(SwerveModule[] modules, SwerveGyroIO gyroIO, boolean sim) {
         super("Drive", DriveStates.REGULAR);
